Imagine you want to go to the pub this evening, but you look too young. The pub staff knows nothing about your real age, so now you need to implement the age checking procedure.
Set the value of 18 to the variable age. Set the condition of age checking to the variable condition: is your age less than 18 or not? If so, you need to write in the console "Sorry, but you are a minor.". If the age is more than or equal to 18, the message is "That's ok. You can enter the pub.".