Write a program that does the following:
It requests the amounts of water, milk, and coffee beans available at the moment and then asks for the number of required coffee cups;
If the coffee machine has the supplies to make enough coffee, the program should print the following: Yes, I can make that amount of coffee.
If the coffee machine can make more than that, the program should output Yes, I can make that amount of coffee (and even N more than that), where N is the number of extra cups.
If you're running short, the program should output No, I can make only N Write a sum function that returns the sum of three variable values, where each is decreased by one. Use variables a, b and c as function parameters.