Write a program that continuously makes coffee for everyone until the shutdown command is provided. Introduce two new options: remaining and exit.
Do not forget that you can overstretch your supplies while making coffee. If the coffee machine doesn't have enough resources to make coffee, the program should output a message saying that it can't make it anymore and state what is missing.

And the last improvement to the program at this step — if users enter buy to buy a cup of coffee but change their mind afterward, allow them to input back to return to the main menu.

Set the following initial supplies: 400 ml of water, 540 ml of milk, 120 g of coffee beans, 9 disposable cups, $550.